+++ src/lib/touchable-highlight/index.ts
@@ -1,0 +1,7 @@
+import { NtTouchableHighlightComponent } from "./nt-touchable-highlight.component";
+import { NtTouchableHighlightModule } from "./nt-touchable-highlight.module";
+
+export {
+    NtTouchableHighlightComponent,
+    NtTouchableHighlightModule,
+}