+++ src/lib/touchable-highlight/utils/index.ts
@@ -1,0 +1,5 @@
+import { roundedRectPath } from './rounded-rect-path';
+
+export {
+    roundedRectPath,
+}