+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
+## [0.6.0] - 2026-09-13
+
+### Added
+- The NtTouchableHighlight component has been added.
+
 ## [0.5.6] - 2026-09-13
 
 ### Fix
-- Stabilization of the Drawer component
-- User interaction fixes
+- Stabilization of the Drawer component.
+- User interaction fixes.
 
 ## [0.5.5] - 2026-09-08